Выполнение домашнего задания. Добавляет модуляцию.
