+++ js/functions.js
@@ -1,22 +1,6 @@
-// // Cтрока короче 20 символов
-// имяФункции('проверяемая строка', 20); // true
-// // Длина строки ровно 18 символов
-// имяФункции('проверяемая строка', 18); // true
-// // Строка длиннее 10 символов
-// имяФункции('проверяемая строка', 10); // false
-
-const stringLengt = (string, length) => {
-  return string.length <= length;
-};
+const stringLengt = (string, length) => string.length <= length;
 stringLengt('проверяемая строка', 20);
 
-// // Строка является палиндромом
-// имяФункции('топот'); // true
-// // Несмотря на разный регистр, тоже палиндром
-// имяФункции('ДовОд'); // true
-// // Это не палиндром
-// имяФункции('Кекс');  // false
-// имяФункции('Лёша на полке клопа нашёл '); // true
 
 const isPalindrom = (unediteSdtring) => {
   const string = unediteSdtring.replaceAll(' ', '').toLowerCase();
@@ -30,15 +14,6 @@ const isPalindrom = (unediteSdtring) => {
 isPalindrom('топот');
 
 
-// имяФункции('2023 год');            // 2023
-// имяФункции('ECMAScript 2022');     // 2022
-// имяФункции('1 кефир, 0.5 батона'); // 105
-// имяФункции('агент 007');           // 7
-// имяФункции('а я томат');           // NaN
-// имяФункции(2023); // 2023
-// имяФункции(-1);   // 1
-// имяФункции(1.5);  // 15
-
 const extractNumber = (unediteSdtring) => {
   const string = unediteSdtring;
   let result = '';
@@ -50,3 +25,5 @@ const extractNumber = (unediteSdtring) => {
   return parseInt(result, 10);
 };
 extractNumber('2023 год');
+
+export {stringLengt, isPalindrom, extractNumber};
